pyproject.toml
[project]
name = "pokeagent-speedrun"
version = "0.1.0"
description = "AI agent that plays Pokémon Emerald using vision-language models"
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
    "flask==3.0.2",
    "flask-socketio==5.3.6",
    "python-socketio==5.11.1",
    "python-engineio==4.9.0",
    "numpy==1.26.4",
    "opencv-python==4.9.0.80",
    "pillow==10.2.0",
    "pygame==2.6.1",
    "requests>=2.32.2",
    "python-dotenv==1.0.1",
    "eventlet==0.35.2",
    "gevent==23.9.1",
    "gevent-websocket==0.10.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "mgba==0.10.2",
    "pyyaml>=5.1",
    "xxhash",
    "more-itertools",
    "pytz>=2020.1",
    "six>=1.14.0",
    "colorama",
    "grpcio>=1.48.2",
    "protobuf!=4.24.0,<5.0.0,>=3.19.6",
    "openai==1.90.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.3.0",
    "cffi>=1.6",
    "pytesseract>=0.3.10",
    "torch==2.10.0",
    "torchvision==0.25.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "beautifulsoup4>=4.14.2",
    "google>=3.0.0",
    "pip>=25.2",
    "mcp>=1.18.0",
    "pyboy==2.5.2",
    "playwright>=1.40.0",
    # Versions tracked to the official Unsloth Gemma4 vision notebook
    # (https://colab.research.google.com/github/unslothai/notebooks/blob/main/nb/Gemma4_(E4B)-Vision.ipynb)
    # which pins datasets==4.3.0 + transformers==5.5.0 and uses
    # trl<0.25 / unsloth>=2026.4.
    "huggingface-hub>=1.5.0",
    "peft>=0.18.0",
    "trl>=0.18.2,<0.25.0",
    "datasets>=4.3.0,<4.4.0",
    "transformers>=5.5.0",
    # No flash-attn — Gemma4 has 7 "global" text layers with head_dim
    # 512 which flash-attn 2 doesn't support. Unsloth + xformers
    # handles those layers via memory_efficient_attention instead.
    "xformers>=0.0.34",
    "liger-kernel>=0.7.0",
    "unsloth>=2026.4.0",
    "unsloth-zoo>=2026.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "pokemon_env", "browser_env", "utils", "server"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.isort]
known-first-party = ["agent", "pokemon_env", "utils", "server"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
